requirements.txt
python-docx>=0.8.11
PyPDF2>=3.0.0
ollama>=0.1.0